chore : validateTheme 파일명 수정 (#97)
checkCalendarDateStyle으로 수정 및 관련 타입명 수정
